app/requirements.txt
streamlit==1.31.0
gspread==5.12.0
oauth2client==4.1.3
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.120.0
Pillow==10.2.0
folium==0.14.0
streamlit-folium==0.16.0
pandas==2.2.0
openpyxl==3.1.2
streamlit-javascript==0.1.5
streamlit-geolocation==0.0.10
